Inventory: change model to be time-zone agnostic
This change makes the inventory model not care about the current time
zone. Previously it mattered a lot which time zone we were in, because
inventory was configured by day. Now that doesn't matter.

This change also includes the ability to suggest other times to the
user, which can help with finding a nearby reservation.